backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
langgraph==0.3.34
openai==1.76.2
google-genai==1.13.0
python-multipart==0.0.9
pandas==2.2.3
XlsxWriter==3.2.0
